+++ client/src/input/Buttons.js
@@ -1,39 +1,18 @@
 import $ from 'jquery'
 import Craft from 'craft'
-import Garnish from 'garnish'
+import { BlockSelector, GarnishBlockSelector } from './BlockSelector'
 
 const _defaults = {
-  $ownerContainer: null,
-  blockTypes: [],
-  groups: [],
-  items: null,
   maxBlocks: 0,
   maxTopBlocks: 0,
   blocks: null
 }
 
-export default Garnish.Base.extend({
-
-  _blockTypes: [],
-  _groups: [],
-  _maxBlocks: 0,
-  _maxTopBlocks: 0,
-
-  init (settings = {}) {
+class Buttons extends BlockSelector {
+  constructor (settings = {}) {
     settings = Object.assign({}, _defaults, settings)
+    super(settings)
 
-    if (settings.items) {
-      this._items = Array.from(settings.items)
-      this._blockTypes = this._items.filter(i => i.getType() === 'blockType')
-      this._groups = this._items.filter(i => i.getType() === 'group')
-    } else {
-      this._blockTypes = Array.from(settings.blockTypes)
-      this._groups = Array.from(settings.groups)
-      this._items = [...this._blockTypes, ...this._groups].sort((a, b) => a.getSortOrder() - b.getSortOrder())
-    }
-
-    this.$ownerContainer = settings.$ownerContainer
-    this._field = settings.field
     this._maxBlocks = settings.maxBlocks | 0
     this._maxTopBlocks = settings.maxTopBlocks | 0
 
@@ -46,18 +25,16 @@ export default Garnish.Base.extend({
     this.$groupButtons = $neo.filter('[data-neo-bn="button.group"]')
 
     if (settings.blocks) {
-      this.updateButtonStates(settings.blocks)
+      this.updateState(settings.blocks)
     }
-
-    this.addListener(this.$blockButtons, 'activate', '@newBlock')
-  },
+  }
 
   _generateButtons () {
     const ownerBlockType = this.$ownerContainer?.hasClass('ni_block')
       ? this.$ownerContainer.attr('class').match(/ni_block--([^\s]+)/)[1]
       : null
     const ungroupChildBlockTypes = ownerBlockType !== null &&
-      !this._field.getBlockTypeByHandle(ownerBlockType).getGroupChildBlockTypes()
+      !this.getField().getBlockTypeByHandle(ownerBlockType).getGroupChildBlockTypes()
     const buttonsHtml = []
     let blockTypesHtml = []
     let currentGroup = null
@@ -95,11 +72,11 @@ export default Garnish.Base.extend({
         if (currentGroup !== null) {
           blockTypesHtml.push(`
             <li>
-              <a${titleAttr} aria-label="${item.getName()}" data-neo-bn="button.addBlock" data-neo-bn-info="${item.getHandle()}">${item.getName()}</a>
+              <a${titleAttr} aria-label="${item.getName()}" data-neo-bn="button.addBlock" ${BlockSelector.BUTTON_INFO}="${item.getHandle()}">${item.getName()}</a>
             </li>`)
         } else {
           buttonsHtml.push(`
-          <button${titleAttr} aria-label="${item.getName()}" class="btn dashed${firstButton ? ' add icon' : ''}" data-neo-bn="button.addBlock" data-neo-bn-info="${item.getHandle()}">
+          <button${titleAttr} aria-label="${item.getName()}" class="btn dashed${firstButton ? ' add icon' : ''}" data-neo-bn="button.addBlock" ${BlockSelector.BUTTON_INFO}="${item.getHandle()}">
             ${item.getName()}
           </button>`)
           firstButton = false
@@ -161,7 +138,7 @@ export default Garnish.Base.extend({
         const titleAttr = item.getDescription() ? ` title="${item.getDescription()}"` : ''
         buttonsHtml.push(`
             <li>
-              <a${titleAttr} aria-label="${item.getName()}" data-neo-bn="button.addBlock" data-neo-bn-info="${item.getHandle()}">
+              <a${titleAttr} aria-label="${item.getName()}" data-neo-bn="button.addBlock" ${BlockSelector.BUTTON_INFO}="${item.getHandle()}">
                 ${item.getName()}
               </a>
             </li>`)
@@ -187,7 +164,7 @@ export default Garnish.Base.extend({
       </div>`)
 
     return $(buttonsHtml.join(''))
-  },
+  }
 
   initUi () {
     $('.menubtn', this.$container).menubtn()
@@ -205,21 +182,9 @@ export default Garnish.Base.extend({
         parent.addClass('hidden')
       }
     }
-  },
+  }
 
-  getBlockTypes () {
-    return Array.from(this._blockTypes)
-  },
-
-  getGroups () {
-    return Array.from(this._groups)
-  },
-
-  getMaxBlocks () {
-    return this._maxBlocks
-  },
-
-  updateButtonStates (blocks = [], additionalCheck = null, block = null) {
+  updateState (blocks = [], additionalCheck = null, block = null) {
     additionalCheck = typeof additionalCheck === 'boolean' ? additionalCheck : true
 
     const that = this
@@ -265,7 +230,7 @@ export default Garnish.Base.extend({
 
       $button.toggleClass('disabled', disabled)
     })
-  },
+  }
 
   updateResponsiveness () {
     this._buttonsContainerWidth ||= this.$buttonsContainer.width()
@@ -273,21 +238,11 @@ export default Garnish.Base.extend({
 
     this.$buttonsContainer.toggleClass('hidden', isMobile)
     this.$menuContainer.toggleClass('hidden', !isMobile)
-  },
+  }
+}
 
-  getBlockTypeByButton ($button) {
-    const btHandle = $button.attr('data-neo-bn-info')
-
-    return this._blockTypes.find(bt => bt.getHandle() === btHandle)
-  },
-
-  '@newBlock' (e) {
-    const $button = $(e.currentTarget)
-    const blockTypeHandle = $button.attr('data-neo-bn-info')
-    const blockType = this._blockTypes.find(bt => bt.getHandle() === blockTypeHandle)
-
-    this.trigger('newBlock', {
-      blockType
-    })
+export default GarnishBlockSelector.extend({
+  init (settings = {}) {
+    this.base(new Buttons(settings))
   }
 })