+++ src/assets/SettingsAsset.php
@@ -3,29 +3,16 @@
 namespace benf\neo\assets;
 
 use benf\neo\elements\Block;
-use benf\neo\events\SetConditionElementTypesEvent;
 use benf\neo\Field;
 use benf\neo\models\BlockType;
 use benf\neo\models\BlockTypeGroup;
 use benf\neo\Plugin as Neo;
 use Craft;
-use craft\commerce\elements\Order;
-use craft\commerce\elements\Product;
-use craft\commerce\elements\Subscription;
-use craft\commerce\elements\Variant;
-use craft\elements\Address;
-use craft\elements\Asset;
-use craft\elements\Category;
-use craft\elements\Entry;
-use craft\elements\Tag;
-use craft\elements\User;
-use craft\helpers\Cp;
 use craft\helpers\Json;
 use craft\helpers\StringHelper;
 use craft\models\FieldLayout;
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
-use yii\base\Event;
 
 /**
  * Class SettingsAsset
@@ -58,11 +45,6 @@ class SettingsAsset extends AssetBundle
      * @since 3.2.0
      */
     public const EVENT_SET_CONDITION_ELEMENT_TYPES = 'setConditionElementTypes';
-
-    /**
-     * @var string[] Supported element types for setting conditions on when block types can be used
-     */
-    private static array $_conditionElementTypes = [];
 
     /**
      * @inheritdoc
@@ -138,15 +120,9 @@ class SettingsAsset extends AssetBundle
      */
     public static function createSettingsJs(Field $field): string
     {
-        $event = new SetConditionElementTypesEvent([
-            'elementTypes' => self::_getSupportedConditionElementTypes(),
-        ]);
-        Event::trigger(self::class, self::EVENT_SET_CONDITION_ELEMENT_TYPES, $event);
-        self::$_conditionElementTypes = $event->elementTypes;
-
         $blockTypes = $field->getBlockTypes();
         $blockTypeGroups = $field->getGroups();
-        [$blockTypeSettingsHtml, $blockTypeSettingsJs] = self::_renderBlockTypeSettings();
+        [$blockTypeSettingsHtml, $blockTypeSettingsJs] = Neo::$plugin->blockTypes->renderBlockTypeSettings();
         $fieldLayoutHtml = Neo::$plugin->blockTypes->renderFieldLayoutDesigner(new FieldLayout(['type' => Block::class]));
 
         $jsSettings = [
@@ -176,7 +152,7 @@ class SettingsAsset extends AssetBundle
         $jsBlockTypes = [];
 
         foreach ($blockTypes as $blockType) {
-            [$blockTypeSettingsHtml, $blockTypeSettingsJs] = self::_renderBlockTypeSettings($blockType);
+            [$settingsHtml, $settingsJs] = Neo::$plugin->blockTypes->renderBlockTypeSettings($blockType);
             $jsBlockTypes[] = [
                 'id' => $blockType->id,
                 'sortOrder' => $blockType->sortOrder,
@@ -185,6 +161,7 @@ class SettingsAsset extends AssetBundle
                 'enabled' => $blockType->enabled,
                 'ignorePermissions' => $blockType->ignorePermissions,
                 'description' => $blockType->description,
+                'iconId' => $blockType->iconId,
                 'minBlocks' => $blockType->minBlocks,
                 'maxBlocks' => $blockType->maxBlocks,
                 'minSiblingBlocks' => $blockType->minSiblingBlocks,
@@ -195,8 +172,8 @@ class SettingsAsset extends AssetBundle
                 'childBlocks' => is_string($blockType->childBlocks) ? Json::decodeIfJson($blockType->childBlocks) : $blockType->childBlocks,
                 'topLevel' => (bool)$blockType->topLevel,
                 'errors' => $blockType->getErrors(),
-                'settingsHtml' => $blockTypeSettingsHtml,
-                'settingsJs' => $blockTypeSettingsJs,
+                'settingsHtml' => $settingsHtml,
+                'settingsJs' => $settingsJs,
                 'fieldLayoutId' => $blockType->fieldLayoutId,
                 'fieldLayoutConfig' => $blockType->getFieldLayout()->getConfig(),
                 'groupId' => $blockType->groupId,
@@ -226,93 +203,5 @@ class SettingsAsset extends AssetBundle
         }
 
         return $jsBlockTypeGroups;
-    }
-
-    /**
-     * @param BlockType|null $blockType
-     * @return array
-     */
-    private static function _renderBlockTypeSettings(?BlockType $blockType = null): array
-    {
-        $view = Craft::$app->getView();
-        $blockTypeId = $blockType?->id ?? '__NEOBLOCKTYPE_ID__';
-        $oldNamespace = $view->getNamespace();
-        $newNamespace = $oldNamespace . '[blockTypes][' . $blockTypeId . ']';
-        $view->setNamespace($newNamespace);
-        $view->startJsBuffer();
-
-        $html = $view->namespaceInputs($view->renderTemplate('neo/block-type-settings', [
-            'blockType' => $blockType,
-            'conditions' => self::_getConditions($blockType),
-        ]));
-
-        $js = $view->clearJsBuffer();
-        $view->setNamespace($oldNamespace);
-
-        return [$html, $js];
-    }
-
-    /**
-     * Gets the condition builder field HTML for a block type.
-     *
-     * @param BlockType|null $blockType
-     * @return string[]
-     */
-    private static function _getConditions(?BlockType $blockType = null): array
-    {
-        $conditionsService = Craft::$app->getConditions();
-        $conditionHtml = [];
-        Neo::$isGeneratingConditionHtml = true;
-
-        foreach (self::$_conditionElementTypes as $elementType) {
-            $condition = !empty($blockType?->conditions) && isset($blockType->conditions[$elementType])
-                ? $conditionsService->createCondition($blockType->conditions[$elementType])
-                : $elementType::createCondition();
-            $condition->mainTag = 'div';
-            $condition->id = 'conditions-' . StringHelper::toKebabCase($elementType);
-            $condition->name = "conditions[$elementType]";
-            $condition->forProjectConfig = true;
-
-            $conditionHtml[$elementType] = Cp::fieldHtml($condition->getBuilderHtml(), [
-                'label' => Craft::t('neo', '{type} Condition', [
-                    'type' => StringHelper::mb_ucwords($elementType::displayName()),
-                ]),
-                'instructions' => Craft::t('neo', 'Only allow this block type to be used on {type} if they match the following rules:', [
-                    'type' => $elementType::pluralLowerDisplayName(),
-                ]),
-            ]);
-        }
-
-        Neo::$isGeneratingConditionHtml = false;
-
-        return $conditionHtml;
-    }
-
-    /**
-     * Get the element types supported by Neo for block type conditionals.
-     *
-     * @return string[]
-     */
-    private static function _getSupportedConditionElementTypes(): array
-    {
-        // In-built Craft element types
-        $elementTypes = [
-            Entry::class,
-            Category::class,
-            Asset::class,
-            User::class,
-            Tag::class,
-            Address::class,
-        ];
-
-        // Craft Commerce element types
-        if (Craft::$app->getPlugins()->isPluginInstalled('commerce')) {
-            $elementTypes[] = Product::class;
-            $elementTypes[] = Variant::class;
-            $elementTypes[] = Order::class;
-            $elementTypes[] = Subscription::class;
-        }
-
-        return $elementTypes;
     }
 }